requirements.txt
ruamel.yaml
httpx
python-dotenv
Pillow
